requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
rumps>=0.4.0
PyQt6>=6.5.0
pynput>=1.7.6
pytesseract>=0.3.10
Pillow>=10.0.0
argostranslate>=1.9.0
